requirements.txt
pandas
openpyxl
Pillow
google-genai
python-dateutil
aiofiles
python-dotenv
